python-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-magic-bin==0.4.14
aiofiles==23.2.1
